Moving to gn build.

+++ src/tools/clang/scripts/package.py
@@ -17,6 +17,7 @@ import tarfile
 
 # Path constants.
 THIS_DIR = os.path.dirname(__file__)
+CHROMIUM_DIR = os.path.abspath(os.path.join(THIS_DIR, '..', '..', '..'))
 THIRD_PARTY_DIR = os.path.join(THIS_DIR, '..', '..', '..', 'third_party')
 LLVM_DIR = os.path.join(THIRD_PARTY_DIR, 'llvm')
 LLVM_BOOTSTRAP_DIR = os.path.join(THIRD_PARTY_DIR, 'llvm-bootstrap')
@@ -24,6 +25,7 @@ LLVM_BOOTSTRAP_INSTALL_DIR = os.path.join(THIRD_PARTY_DIR,
                                           'llvm-bootstrap-install')
 LLVM_BUILD_DIR = os.path.join(THIRD_PARTY_DIR, 'llvm-build')
 LLVM_RELEASE_DIR = os.path.join(LLVM_BUILD_DIR, 'Release+Asserts')
+LLVM_LTO_GOLD_PLUGIN_DIR = os.path.join(THIRD_PARTY_DIR, 'llvm-lto-gold-plugin')
 STAMP_FILE = os.path.join(LLVM_BUILD_DIR, 'cr_build_revision')
 
 
@@ -57,16 +59,92 @@ def PrintTarProgress(tarinfo):
   return tarinfo
 
 
-def main():
-  if sys.platform == 'win32':
-    try:
-      subprocess.check_output(['grep', '--help'], shell=True)
-    except subprocess.CalledProcessError:
-      print 'Add gnuwin32 to your PATH, then try again.'
-      return 1
+def GetExpectedStamp():
+  rev_cmd = [sys.executable, os.path.join(THIS_DIR, 'update.py'),
+             '--print-revision']
+  return subprocess.check_output(rev_cmd).rstrip()
 
+
+def GetGsutilPath():
+  if not 'find_depot_tools' in sys.modules:
+    sys.path.insert(0, os.path.join(CHROMIUM_DIR, 'build'))
+    global find_depot_tools
+    import find_depot_tools
+  depot_path = find_depot_tools.add_depot_tools_to_path()
+  if depot_path is None:
+    print ('depot_tools are not found in PATH. '
+           'Follow the instructions in this document '
+           'http://dev.chromium.org/developers/how-tos/install-depot-tools'
+           ' to install depot_tools and then try again.')
+    sys.exit(1)
+  gsutil_path = os.path.join(depot_path, 'gsutil.py')
+  return gsutil_path
+
+
+def RunGsutil(args):
+  return subprocess.call([sys.executable, GetGsutilPath()] + args)
+
+
+def GsutilArchiveExists(archive_name, platform):
+  gsutil_args = ['-q', 'stat',
+                 'gs://chromium-browser-clang/%s/%s.tgz' %
+                 (platform, archive_name)]
+  return RunGsutil(gsutil_args) == 0
+
+
+def MaybeUpload(args, archive_name, platform):
+  # We don't want to rewrite the file, if it already exists on the server,
+  # so -n option to gsutil is used. It will warn, if the upload was aborted.
+  gsutil_args = ['cp', '-n', '-a', 'public-read',
+                  '%s.tgz' % archive_name,
+                  'gs://chromium-browser-clang/%s/%s.tgz' %
+                 (platform, archive_name)]
+  if args.upload:
+    print 'Uploading %s to Google Cloud Storage...' % archive_name
+    exit_code = RunGsutil(gsutil_args)
+    if exit_code != 0:
+      print "gsutil failed, exit_code: %s" % exit_code
+      os.exit(exit_code)
+  else:
+    print 'To upload, run:'
+    print ('gsutil %s' % ' '.join(gsutil_args))
+
+
+def main():
   parser = argparse.ArgumentParser(description='build and package clang')
+  parser.add_argument('--upload', action='store_true',
+                      help='Upload the target archive to Google Cloud Storage.')
   args = parser.parse_args()
+
+  # Check that the script is not going to upload a toolchain built from HEAD.
+  use_head_revision = 'LLVM_FORCE_HEAD_REVISION' in os.environ
+  if args.upload and use_head_revision:
+    print ("--upload and LLVM_FORCE_HEAD_REVISION could not be used "
+           "at the same time.")
+    return 1
+
+  expected_stamp = GetExpectedStamp()
+  pdir = 'clang-' + expected_stamp
+  golddir = 'llvmgold-' + expected_stamp
+  print pdir
+
+  if sys.platform == 'darwin':
+    platform = 'Mac'
+  elif sys.platform == 'win32':
+    platform = 'Win'
+  else:
+    platform = 'Linux_x64'
+
+  # Check if Google Cloud Storage already has the artifacts we want to build.
+  if (args.upload and GsutilArchiveExists(pdir, platform) and
+      not sys.platform.startswith('linux') or
+      GsutilArchiveExists(golddir, platform)):
+    print ('Desired toolchain revision %s is already available '
+           'in Google Cloud Storage:') % expected_stamp
+    print 'gs://chromium-browser-clang/%s/%s.tgz' % (platform, pdir)
+    if sys.platform.startswith('linux'):
+      print 'gs://chromium-browser-clang/%s/%s.tgz' % (platform, golddir)
+    return 0
 
   with open('buildlog.txt', 'w') as log:
     Tee('Diff in llvm:\n', log)
@@ -90,11 +168,6 @@ def main():
            log, fail_hard=False)
     TeeCmd(['svn', 'diff', os.path.join(LLVM_DIR, 'projects', 'libcxx')],
            log, fail_hard=False)
-    Tee('Diff in llvm/projects/libcxxabi:\n', log)
-    TeeCmd(['svn', 'stat', os.path.join(LLVM_DIR, 'projects', 'libcxxabi')],
-           log, fail_hard=False)
-    TeeCmd(['svn', 'diff', os.path.join(LLVM_DIR, 'projects', 'libcxxabi')],
-           log, fail_hard=False)
 
     Tee('Starting build\n', log)
 
@@ -103,19 +176,26 @@ def main():
     shutil.rmtree(LLVM_BOOTSTRAP_INSTALL_DIR, ignore_errors=True)
     shutil.rmtree(LLVM_BUILD_DIR, ignore_errors=True)
 
+    opt_flags = []
+    if sys.platform.startswith('linux'):
+      opt_flags += ['--lto-gold-plugin']
     build_cmd = [sys.executable, os.path.join(THIS_DIR, 'update.py'),
-                 '--bootstrap', '--force-local-build', '--run-tests']
+                 '--bootstrap', '--force-local-build',
+                 '--run-tests'] + opt_flags
     TeeCmd(build_cmd, log)
 
   stamp = open(STAMP_FILE).read().rstrip()
-  pdir = 'clang-' + stamp
-  print pdir
+  if stamp != expected_stamp:
+    print 'Actual stamp (%s) != expected stamp (%s).' % (stamp, expected_stamp)
+    return 1
+
   shutil.rmtree(pdir, ignore_errors=True)
 
   # Copy a whitelist of files to the directory we're going to tar up.
   # This supports the same patterns that the fnmatch module understands.
   exe_ext = '.exe' if sys.platform == 'win32' else ''
   want = ['bin/llvm-symbolizer' + exe_ext,
+          'bin/sancov' + exe_ext,
           'lib/clang/*/asan_blacklist.txt',
           'lib/clang/*/cfi_blacklist.txt',
           # Copy built-in headers (lib/clang/3.x.y/include).
@@ -131,15 +211,15 @@ def main():
                  'lib/libBlinkGCPlugin.' + so_ext,
                  ])
   if sys.platform == 'darwin':
-    want.extend(['bin/libc++.1.dylib',
-                 # Copy only the OSX (ASan and profile) and iossim (ASan)
-                 # runtime libraries:
+    want.extend([# Copy only the OSX and iossim (ASan and profile) runtime
+                 # libraries:
                  'lib/clang/*/lib/darwin/*asan_osx*',
                  'lib/clang/*/lib/darwin/*asan_iossim*',
                  'lib/clang/*/lib/darwin/*profile_osx*',
+                 'lib/clang/*/lib/darwin/*profile_iossim*',
                  ])
   elif sys.platform.startswith('linux'):
-    # Copy the stdlibc++.so.6 we linked Clang against so it can run.
+    # Copy the libstdc++.so.6 we linked Clang against so it can run.
     want.append('lib/libstdc++.so.6')
     # Copy only
     # lib/clang/*/lib/linux/libclang_rt.{[atm]san,san,ubsan,profile}-*.a ,
@@ -180,9 +260,9 @@ def main():
   if sys.platform != 'win32':
     os.symlink('clang', os.path.join(pdir, 'bin', 'clang++'))
     os.symlink('clang', os.path.join(pdir, 'bin', 'clang-cl'))
+
+  # Copy libc++ headers.
   if sys.platform == 'darwin':
-    os.symlink('libc++.1.dylib', os.path.join(pdir, 'bin', 'libc++.dylib'))
-    # Also copy libc++ headers.
     shutil.copytree(os.path.join(LLVM_BOOTSTRAP_INSTALL_DIR, 'include', 'c++'),
                     os.path.join(pdir, 'include', 'c++'))
 
@@ -197,30 +277,29 @@ def main():
     for entry in tar_entries:
       tar.add(os.path.join(pdir, entry), arcname=entry, filter=PrintTarProgress)
 
-  if sys.platform == 'darwin':
-    platform = 'Mac'
-  elif sys.platform == 'win32':
-    platform = 'Win'
-  else:
-    platform = 'Linux_x64'
-
-  print 'To upload, run:'
-  print ('gsutil cp -a public-read %s.tgz '
-         'gs://chromium-browser-clang/%s/%s.tgz') % (pdir, platform, pdir)
+  MaybeUpload(args, pdir, platform)
 
   # Zip up gold plugin on Linux.
   if sys.platform.startswith('linux'):
-    golddir = 'llvmgold-' + stamp
     shutil.rmtree(golddir, ignore_errors=True)
     os.makedirs(os.path.join(golddir, 'lib'))
-    shutil.copy(os.path.join(LLVM_RELEASE_DIR, 'lib', 'LLVMgold.so'),
+    shutil.copy(os.path.join(LLVM_LTO_GOLD_PLUGIN_DIR, 'lib', 'LLVMgold.so'),
                 os.path.join(golddir, 'lib'))
     with tarfile.open(golddir + '.tgz', 'w:gz') as tar:
       tar.add(os.path.join(golddir, 'lib'), arcname='lib',
               filter=PrintTarProgress)
-    print ('gsutil cp -a public-read %s.tgz '
-           'gs://chromium-browser-clang/%s/%s.tgz') % (golddir, platform,
-                                                       golddir)
+    MaybeUpload(args, golddir, platform)
+
+  # Zip up llvm-objdump for sanitizer coverage.
+  objdumpdir = 'llvmobjdump-' + stamp
+  shutil.rmtree(objdumpdir, ignore_errors=True)
+  os.makedirs(os.path.join(objdumpdir, 'bin'))
+  shutil.copy(os.path.join(LLVM_RELEASE_DIR, 'bin', 'llvm-objdump' + exe_ext),
+              os.path.join(objdumpdir, 'bin'))
+  with tarfile.open(objdumpdir + '.tgz', 'w:gz') as tar:
+    tar.add(os.path.join(objdumpdir, 'bin'), arcname='bin',
+            filter=PrintTarProgress)
+  MaybeUpload(args, objdumpdir, platform)
 
   # FIXME: Warn if the file already exists on the server.
